Fix punctuation
https://jakubmarian.com/each-others-vs-each-others-in-english/

This is quite logical. The possessive form in English is formed by adding ’s at the end of a noun, unless it is a plural noun, in which case we write just an apostrophe, e.g. “these teachers’ books” (not “these teachers’s books”). This rules out each others, as the possessive apostrophe must be there.

In the case of “each other”, “other” is in the singular because it follows “each”—you wouldn’t say “each teachers” instead of “each teacher”, would you… By adding the possessive ’s, we get the correct form each other’s.