Move all strings I can find into yml files (#36)
This moves all of the strings into yml files so that they can potentially be updated without seeing the html directly. This uses the liquid tags to basically index into the yml and bring in the string tables from yml. In this way we can potentially have things like en/jp strings under different folders in the futue.